Move unicode.py from lib/ to lib/utils/
